Define the Absolute/Relative Orientation Sensor sensor types in this spec
Contrary to other specifications such as Accelerometer and Gyroscope, this
specification did not define a sensor type for each of the concrete
Orientation Sensors -- there was only a definition for Orientation Sensor
itself.

The concept of inheritance between concepts is not really defined, so we
should have sensor type definitions for absolute and relative orientation
sensors explicitly written out as required by
https://w3c.github.io/sensors/#model-sensor-type, especially to fix #77, as
some automation concepts are hooked to the "sensor type" concept.