editorial: Move the rotation matrix creation algorithm to a separate operation
This will make it easier to invoke from the Automation section in an
upcoming commit.

The "convert a quaternion to rotation matrix" algorithm performs the same
steps that were part of `OrientationSensor.populateMatrix()` and returns the
matrix as a 16-element list that is used by the latter.

There are no changes to the algorithm itself.

Drive-by improvements to populateMatrix():
- Stop checking `targetMatrix`'s type, this is already done by Web IDL.
- Refer to `this` with the right notation.